requirements.txt
python>=3.6
django>=3.2
djangorestframework
drf-yasg
pytest